Add subtitles
This parses the stream url for subtitle links, which are on the form

  http://url.com/video?cap&c1_file=url_to_sub1.vtt&c1_label=language_1&c2_file=...

For now, only parse the link(s). In most cases, it seems to be only one
language (Swedish for the most part, English in some unusual cases).

In most cases, there is not indication of language in the subtitle url,
which makes it, in the cases of multiple languages available, no way
to figure out which subtitle is in which language, except for trying them
and recognizing the language. However, since this is an unusual case,
we ignore that for now.