Bugfix Picasa resolver
Previously, the Picasa resolver tried parsing the html-page for links to
the available streams. This breaks when the album contains more than one
video, in which case the first one was always returned.

Instead, use the JSON-api. This involves doing some regexp magic with the
url to parse user, album, etc.