Handle leading and trailing whitespaces in subtitles
It was discovered that some stream urls could be returned with leading
whitespaces, causing the plugin to fail with the warning

    WARNING: CreateLoader - unsupported protocol( https) in  https://...

Note the leading space. This was not observed for subtitles, but adjust
the regexp to avoid capturing leading and trailing whitespaces to make
sure we don't run into this for subtitles.