Yet again fix vkpass resolver
In a previous commit it was discovered that the
vkpass pages sometimes does not contain the video
urls but a link to a page that does. Then it turned
out that sometimes a page is show not in clear text
but binary coded. The solution to these involved recursion,
and to make sure we do not stop in an infinite recursion
callchain, we only allowed one recursive call. It now
turns out that the redirect page can be binary decoded,
which means we'll have to do even one more layer of
recursion. So allow a few more recursions (ten chosen
completely arbitrary) to make sure we catch these cases
as well.